+++ vehicle/OVMS.V3/main/ovms_config.cpp
@@ -354,6 +354,12 @@ void OvmsConfig::upgrade()
       DeleteInstance("server.v2","password");
       }
     }
+    
+  // Migrate vehicle ID VWUP to VWUP.T26
+  if (GetParamValue("auto", "vehicle.type") == "VWUP")
+    {
+    SetParamValue("auto", "vehicle.type", "VWUP.T26");
+    }
 
   // Done, set config version:
   SetParamValueInt("module", "cfgversion", 2020053100);
